+++ packages/core/src/uniquiness/factory.ts
@@ -1,3 +1,4 @@
+import { inspect } from "util";
 import { logger } from "../logger";
 import { AliveJobUniqueness } from "./alive-job-uniqueness";
 import { FixedWindowUniqueness } from "./fixed-window-uniqueness";
@@ -21,7 +22,7 @@ export class UniquenessFactory {
    */
   static create(config: UniquenessConfig): Uniqueness {
     const Ctor = availableUniqueness[config.type] as UniquenessConstructor | undefined;
-    logger("Core").debug(`Creating uniqueness strategy: ${config.type}. Config: ${JSON.stringify(config)}`);
+    logger("Core").debug(`Creating uniqueness strategy: ${config.type}. Config: ${inspect(config)}`);
     if (!Ctor) {
       throw new Error(`Unknown uniqueness strategy: ${config.type}`);
     }